+++ src/main/java/antoniogiovanni/marchese/TEAM4BW5/model/Comune.java
@@ -13,7 +13,7 @@ import lombok.*;
 @ToString
 public class Comune {
     @Id
-    private int progressivoDelComune;
+    private long progressivoDelComune;
 
     private int codiceProvincia;
     private String denominazione;
