+++ src/main/java/antoniogiovanni/marchese/TEAM4BW5/payloads/ComuneDTO.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class ComuneDTO {
     int codiceProvincia; String denominazione; Provincia provincia;
 
-    public ComuneDTO(int progressivoDelComune) {
+    public ComuneDTO(long progressivoDelComune) {
     }
 }
